adding confirm alert #1626

+++ dcm4chee-arc-ui2/src/app/monitoring/external-retrieve/external-retrieve.component.ts
@@ -365,27 +365,36 @@ export class ExternalRetrieveComponent implements OnInit,OnDestroy {
         });
     }
     deleteBatchedTask(batchedTask){
-        if(batchedTask.properties.batchID){
-            let filter = Object.assign({},this.filterObject);
-            filter["batchID"] = batchedTask.properties.batchID;
-            this.service.deleteAll(filter).subscribe((res)=>{
-                this.mainservice.setMessage({
-                    'title': 'Info',
-                    'text': res.deleted + ' tasks deleted successfully!',
-                    'status': 'info'
-                });
-                this.cfpLoadingBar.complete();
-            }, (err) => {
-                this.cfpLoadingBar.complete();
-                this.httpErrorHandler.handleError(err);
-            });
-        }else{
-            this.mainservice.setMessage({
-                'title': 'Error',
-                'text': 'Batch ID not found!',
-                'status': 'error'
-            });
-        }
+        this.confirm({
+            content: 'Are you sure you want to delete all tasks to this batch?'
+        }).subscribe(ok=>{
+            if(ok){
+                if(batchedTask.properties.batchID){
+                    let filter = Object.assign({},this.filterObject);
+                    filter["batchID"] = batchedTask.properties.batchID;
+                    delete filter["limit"];
+                    delete filter["offset"];
+                    this.service.deleteAll(filter).subscribe((res)=>{
+                        this.mainservice.setMessage({
+                            'title': 'Info',
+                            'text': res.deleted + ' tasks deleted successfully!',
+                            'status': 'info'
+                        });
+                        this.cfpLoadingBar.complete();
+                        this.getTasks(0)
+                    }, (err) => {
+                        this.cfpLoadingBar.complete();
+                        this.httpErrorHandler.handleError(err);
+                    });
+                }else{
+                    this.mainservice.setMessage({
+                        'title': 'Error',
+                        'text': 'Batch ID not found!',
+                        'status': 'error'
+                    });
+                }
+            }
+        });
     }
     delete(match){
         let $this = this;
